blog/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.1.1
cssselect==1.1.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
joblib==0.17.0
lockfile==0.12.2
lxml==4.6.1
markdown2==2.3.10
msgpack==0.6.2
nltk==3.5
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pyquery==1.4.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
readtime==1.1.1
regex==2020.10.23
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.1
textblob==0.15.3
tqdm==4.51.0
urllib3==1.25.8
webencodings==0.5.1
